# faster compareText