--- a/Data/Text.hs
+++ b/Data/Text.hs
@@ -410,17 +410,20 @@ textDataType = mkDataType "Data.Text.Text" [packConstr]
 
 -- | /O(n)/ Compare two 'Text' values lexicographically.
 compareText :: Text -> Text -> Ordering
-compareText ta@(Text _arrA _offA lenA) tb@(Text _arrB _offB lenB)
-    | lenA == 0 && lenB == 0 = EQ
-    | otherwise              = go 0 0
+compareText (Text arrA offA lenA) (Text arrB offB lenB)
+    | lenA == 0 || lenB == 0 = compare lenA lenB
+    | otherwise              = go offA offB
   where
+    ea = offA + lenA
+    eb = offB + lenB
     go !i !j
-        | i >= lenA || j >= lenB = compare lenA lenB
-        | a < b                  = LT
-        | a > b                  = GT
-        | otherwise              = go (i+di) (j+dj)
-      where Iter a di = iter ta i
-            Iter b dj = iter tb j
+        | i >= ea || j >= eb = compare lenA lenB
+        | a < b              = LT
+        | a > b              = GT
+        | otherwise          = go (i+1) (j+1)
+      where
+      a = A.unsafeIndex arrA i
+      b = A.unsafeIndex arrB j
 
 -- -----------------------------------------------------------------------------
 -- * Conversion to/from 'Text'