we don't need to iterate over whole codepoints and decode them into Char's
it is ok to go word by word; this works also for 2-word codepoints
this saves ~10% of time, which is important for sorting/Map Text
benchmark:

```
before:

benchmarking Programs/Sort/Text
time                 591.7 ms   (577.7 ms .. 600.9 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 594.2 ms   (591.8 ms .. 595.8 ms)
std dev              2.352 ms   (0.0 s .. 2.713 ms)
variance introduced by outliers: 19% (moderately inflated)

after:

benchmarking Programs/Sort/Text
time                 531.2 ms   (527.8 ms .. 533.7 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 533.2 ms   (532.3 ms .. 533.7 ms)
std dev              855.6 μs   (0.0 s .. 857.6 μs)
variance introduced by outliers: 19% (moderately inflated)
```