@kuk0 2017-11-03
the improvement is actually 25-50% (when comparing two equal strings - to decide whether x<x we need to compare all characters)
note that `Programs/Sort` above benchmarks the whole program, which consists of reading the data, splitting them into lines and sorting the lines and i only improved sorting

```
benchmarking Pure/compareText/Text+tiny
time                 82.79 ns   (82.76 ns .. 82.83 ns)
time                 60.84 ns   (60.79 ns .. 60.93 ns)
benchmarking Pure/compareText/Text+ascii-small
time                 588.1 μs   (588.0 μs .. 588.3 μs)
time                 293.9 μs   (293.8 μs .. 294.0 μs)
benchmarking Pure/compareText/Text+ascii
time                 510.9 ms   (510.5 ms .. 511.3 ms)
time                 257.1 ms   (256.4 ms .. 257.6 ms)
benchmarking Pure/compareText/Text+english
time                 34.14 ms   (34.13 ms .. 34.15 ms)
time                 17.19 ms   (17.18 ms .. 17.20 ms)
benchmarking Pure/compareText/Text+russian
time                 54.49 μs   (54.48 μs .. 54.51 μs)
time                 27.26 μs   (27.26 μs .. 27.26 μs)
benchmarking Pure/compareText/Text+japanese
time                 37.88 μs   (37.87 μs .. 37.89 μs)
time                 18.82 μs   (18.81 μs .. 18.82 μs)
```