@hvr 2019-06-23
@kuk0

> it is ok to go word by word; this works also for 2-word codepoints

I'm not convinced this is actually sound

For instance, consider the code-points `U+E001` and `U+10001`

Their UTF16 encoding into 16 bit words would be

```
0xe001
0xd800 0xdc01
```

respectively; and yet if you compare them lexicographically as 16-bit words, you'd end up determining that `U+E001 > U+10001` which is obviously not correct.

Am I missing something?