@llelf 2019-06-23
@hvr you are not missing anything, it is wrong.

However there *actually exists* a way to do it, using the additional lookup table, see Unicode core specification, p5.17.   But I’m not quite sure it will be worth it.

This is one another case where UTF-8 is better (because doing it in UTF-8 is straightforward) and why UTF-16 in `Text` was a huge mistake.