@hvr 2019-06-23
@llelf yeah, I suspect the performance improvement will likely be rather modest for the added complexity of using the lookup tables...