@Lysxia 2021-05-12
Fixing this seems like a rather involved affair, so it seems safe to close this.